+++ beets/ui/__init__.py
@@ -366,7 +366,7 @@ def colorize(color, text):
     else:
         return text
 
-def _colordiff(a, b, highlight='red', second_highlight='lightgray'):
+def _colordiff(a, b, highlight='red', minor_highlight='lightgray'):
     """Given two values, return the same pair of strings except with
     their differences highlighted in the specified color. Strings are
     highlighted intelligently to show differences; other values are
@@ -407,7 +407,7 @@ def _colordiff(a, b, highlight='red', second_highlight='lightgray'):
             if a[a_start:a_end].lower() != b[b_start:b_end].lower():
                 color = highlight
             else:
-                color = second_highlight
+                color = minor_highlight
             a_out.append(colorize(color, a[a_start:a_end]))
             b_out.append(colorize(color, b[b_start:b_end]))
         else: